test(e2e): add visitAllLinksAndGetResponseInfo util, use func in login & dashboard tests
