+++ test/e2e/login-page.test.ts
@@ -1,5 +1,10 @@
 import { test, expect } from '@playwright/test';
-import { testTitleWithTags, getAllTestIds, doLogin } from './utils';
+import {
+  testTitleWithTags,
+  getAllTestIds,
+  doLogin,
+  visitAllLinksAndGetResponseInfo,
+} from './utils';
 
 test.describe('LoginPage', () => {
   test.beforeEach(async ({ page }) => {
@@ -108,41 +113,10 @@ test.describe('LoginPage', () => {
       'primary',
     ]),
     async ({ page, context }) => {
-      const linkLocators = await page.getByRole('link').all();
-      const visitedLinks = new Set();
-      for (const link of linkLocators) {
-        const href = await link.getAttribute('href');
-        console.info('Found link', href);
-        const isVisited = visitedLinks.has(href);
-        if (isVisited) {
-          console.info('Skipping already visited link', href);
-        } else if (href && !href.startsWith('mailto:')) {
-          const newPage = await context.newPage();
-          await newPage.bringToFront();
-          const response = await newPage.goto(href, {
-            waitUntil: 'commit',
-          });
-          // Add link to visited links
-          visitedLinks.add(href);
-          if (!response) {
-            console.error('No response from', href);
-            continue;
-          }
-
-          console.info(`Response from ${href}`, {
-            url: response.url(),
-            ok: response.ok(),
-            status: response.status(),
-            headers: response.headers(),
-          });
-          const status = response.status();
-          expect(status).toBe(200);
-          // close newly opened page (tab) to avoid OOM issues
-          await newPage.close();
-          // bring root page back into focus
-          await page.bringToFront();
-        }
-      }
+      const linkInfos = await visitAllLinksAndGetResponseInfo(page, context);
+      const statuses = linkInfos.map((l) => l.status);
+      const isStatusOKForAllLinks = statuses.every((s) => s === 200);
+      expect(isStatusOKForAllLinks).toBe(true);
     },
   );
 });
